[android] Completely location in background
